backend/requirements.txt
fastapi[security]
uvicorn
sqlalchemy
aiomysql
cryptography
passlib[bcrypt]
pydantic[email]
bcrypt
aiomysql
pyjwt
python-dotenv
alembic
